Select all done

+++ components/delete-various-transaction-button.tsx
@@ -3,6 +3,7 @@ import { useRouterRefresh } from '../hooks/hooks';
 import { TrashIcon } from '@heroicons/react/solid';
 import { useToastContext } from '../lib/toast-context';
 import { BASE_URL } from '../types/common';
+import Swal from 'sweetalert2'
 
 interface DeleteTruckTransactionButtonProps {
   transactionId: string;
@@ -36,7 +37,25 @@ export default function DeleteVariousTransactionButton({
       } h-7`}
       onClick={() => {
         if (!disabled) {
-          deleteTruckTransaction();
+          Swal.fire({
+            title: 'Yakin akan menghapus transaksi ini?',
+            text: "Transaksi ini tidak akan bisa dikembalikan setelah dihapus",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: 'gray',
+            confirmButtonText: 'Ya, tolong dihapus!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              deleteTruckTransaction();
+              Swal.fire(
+                'Terhapus!',
+                'Transaksi ini telah berhasil dihapus',
+                'success'
+              )
+            }
+          })
+          
         }
       }}
     />
